djangogramm/requirements.txt
Django==4.1.2

djangorestframework==3.14.0
markdown==3.4.1
django-filter==22.1
djangorestframework-jwt==1.11.0
sentry-sdk==1.9.10

django-model-utils==4.2.0
django-extensions==3.2.1
django-debug-toolbar==3.7.0
django-admin==2.0.1
Pillow==9.2.0
django-taggit==3.0.0

waitress==2.1.2

six==1.16.0

django-cloudinary-storage==0.3.0